Fix bug removing functions

+++ src/city_gen_model/progression/Progression.java
@@ -2,7 +2,7 @@ package city_gen_model.progression;
 
 public interface Progression {
 
-    double getNextValue(int currentValue);
+    double getNextValue(double currentValue);
 
     double getRateOfChange();
 }
